Fix clause for parsing response time

+++ cgi-bin/TestPing.py
@@ -4,7 +4,7 @@ def printPingTime(data):
 	listPing = []
 	for linha in data:
 		if("TTL" in linha):
-			linhaParts = linha.split("tempo=")[1].split(" ")
+			linhaParts = linha.split("time=")[1].split(" ")
 			listPing.append(str(linhaParts[0]))
 		elif("Request timed out" in linha):
 			listPing.append("Lost Package")
@@ -13,8 +13,11 @@ def printPingTime(data):
 		
 	return listPing
 
-def getPing():			
-	os.remove('ping.txt')
+def getPing():		
+
+	if(os.path.exists('ping.txt')):
+		os.remove('ping.txt')
+		
 	os.system("ping google.com >> ping.txt")
 	data = open('ping.txt')
 	return data
